+++ ReleaseNotes.md
@@ -6,9 +6,11 @@
 - Added Hungarian translation. Thanks to StaK_1980!
 - Added amount of HP restored per heal on ship consumables.
 - Added shell trajectory chart.
+- Added concealment and spotting values for planes.
 
 ## Changes
 - Increased spacing between rows in the ship statistic for better readability.
+- Improved stats layout.
 
 ## Bugfixes
 - Fixed Hidden Menace and Damage Control Party Modification 1 not changing stats of damage control party reload for CV